Update: 포인트 수치 오류

+++ Friday Diary/Assets/Scripts/Backend/Backend.cs
@@ -147,7 +147,7 @@ public class Backend : MonoBehaviour
 
     
     // 나무 구매
-    public void CreateObject(string emotion, int cost, Tree tree, Action<string> onSuccess){
+    public void CreateObject(string emotion, double cost, Tree tree, Action<string> onSuccess){
         
         Dictionary<string, string> data = new Dictionary<string, string>();
         data.Add("id",GameManager.i.GetUser().GetId());
